vh: sched: Remove PMU poll feature
The PMU poll feature attempts to poll perf counters to determine if the CPU
is stalled by off-chip accesses, in which case the CPU frequency can be
capped to (apparently) save power.

As it stands, this feature is disabled by default, and it's unclear when
it's ever used in Android. The code quality is poor (`mon_idx` is a stack
variable that's used uninitialized in two places), and racy as well since
statistics are updated on a per-CPU basis out of sync with the poller in
cpufreq_gov. Not to mention, a kthread work is used for polling every 10 ms
instead of just using a kthread on its own, which adds undue overhead.

Since the code is awful and the feature is unused as far as one can tell,
remove it outright. The concept itself is flawed since all CPUs in a
cluster share the same clock source; i.e., frequency cannot be capped on a
per-CPU basis, so the power-saving benefits of this feature are dubious.

Note that the procfs nodes must be kept in order to keep libperfmgr happy,
otherwise libperfmgr constantly complains about the missing procfs files.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
(cherry picked from commit 6642503903c55c5378b365e82b23f36b659621d1)